+++ netbox_catalyst_center/urls.py
@@ -7,6 +7,7 @@ from django.urls import path
 from .views import (
     CatalystCenterSettingsView,
     ImportDevicesView,
+    ImportPageView,
     SearchDevicesView,
     SyncDeviceFromDNACView,
     TestConnectionView,
@@ -14,6 +15,7 @@ from .views import (
 
 urlpatterns = [
     path("settings/", CatalystCenterSettingsView.as_view(), name="settings"),
+    path("import/", ImportPageView.as_view(), name="import_page"),
     path("test-connection/", TestConnectionView.as_view(), name="test_connection"),
     path("sync-device/<int:pk>/", SyncDeviceFromDNACView.as_view(), name="sync_device"),
     path("search-devices/", SearchDevicesView.as_view(), name="search_devices"),